+++ src/app/modules/select-columns/select-columns.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject } from '@angular/core';
 import { ColumnDescription } from '../../types/column.description';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA, MatLegacyDialogRef as MatDialogRef } from '@angular/material/legacy-dialog';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';